eventmanager: Introduce EventManager and implement listener add/remove
Introduce the EventManager singleton to manage cross object events in
the engine. This EventManager will allow any object to add or remove
listeners for any arbitrary event ID, where the listener is a member
function of the signature `void foo(IEventData)` which has been
typedef'd to EventListner. This makes use of std::function to store a
function object created using std::bind. When a listener is added for a
specific event ID, it will be invoked when that event is triggered and
have the event specific IEventData subclass passed to it.

This patch introduces the EventManager singleton, a basic
EventData_EntityCreate event, and the add and remove listener calls. A
future patch will introduce the implementation for the Queue, Trigger,
and Abort methods.