include: interfaces: Introduce ISystem interface
Introduce an ISystem interface class. Unsurprisingly, his will be used
for "systems", which are top level blocks of code that will be updated
every cycle of the game engine. Examples of systems are physics and
rendering, which are both special cases, but generic systems will also
be made available as well that can be added and will tick along with
everything else.